refactor(server): lift the aggregate column parser below both ORMs
Both twenty-orm and twenty-orm-v2 reached up into the GraphQL layer for
extractColumnNamesFromAggregateExpression to work out which columns a
select touches. Move it to src/utils so the dependency points down from
all three call sites, and move its spec with it.
