+++ packages/twenty-server/src/utils/__tests__/extract-column-names-from-aggregate-expression.util.spec.ts
@@ -1,0 +1,57 @@
+import { extractColumnNamesFromAggregateExpression } from 'src/utils/extract-column-names-from-aggregate-expression.util';
+
+describe('extractColumnNamesFromAggregateExpression', () => {
+  it('should extract column names from CONCAT expression', () => {
+    const selection =
+      'CASE WHEN COUNT(*) = 0 THEN NULL ELSE COUNT(*) - COUNT(NULLIF(CONCAT("firstName","lastName")) END, \'\')';
+    const result = extractColumnNamesFromAggregateExpression(selection);
+
+    expect(result).toEqual(['firstName', 'lastName']);
+  });
+
+  it('should extract column names from CONCAT expression - 2', () => {
+    const selection =
+      'CASE WHEN COUNT(*) = 0 THEN NULL ELSE COUNT(*) - COUNT(NULLIF(CONCAT("firstName")) END, \'\')';
+    const result = extractColumnNamesFromAggregateExpression(selection);
+
+    expect(result).toEqual(['firstName']);
+  });
+
+  it('should extract column names from CONCAT expression - 3', () => {
+    const selection =
+      'CASE WHEN COUNT(*) = 0 THEN NULL ELSE COUNT(*) - COUNT(NULLIF(CONCAT("firstName","lastName","nickName")) END, \'\')';
+    const result = extractColumnNamesFromAggregateExpression(selection);
+
+    expect(result).toEqual(['firstName', 'lastName', 'nickName']);
+  });
+
+  it('should extract column name from non-CONCAT expression', () => {
+    const selection =
+      'CASE WHEN COUNT(*) = 0 THEN NULL ELSE COUNT("firstName") END';
+    const result = extractColumnNamesFromAggregateExpression(selection);
+
+    expect(result).toEqual(['firstName']);
+  });
+
+  it('should extract column name from aggregate expression', () => {
+    const selection = 'AVG("amount")';
+    const result = extractColumnNamesFromAggregateExpression(selection);
+
+    expect(result).toEqual(['amount']);
+  });
+
+  it('should return null when no column names found', () => {
+    const selection = 'COUNT(*)';
+    const result = extractColumnNamesFromAggregateExpression(selection);
+
+    expect(result).toBeNull();
+  });
+
+  it('should extract column name from boolean expression', () => {
+    const selection =
+      'CASE WHEN "isActive"::boolean = TRUE THEN 1 ELSE NULL END';
+    const result = extractColumnNamesFromAggregateExpression(selection);
+
+    expect(result).toEqual(['isActive']);
+  });
+});
